move config related to config folder and add set storage group related
