+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/config/ConfigExecution.java
@@ -17,9 +17,12 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.execution;
+package org.apache.iotdb.db.mpp.execution.config;
 
 import org.apache.iotdb.db.mpp.common.MPPQueryContext;
+import org.apache.iotdb.db.mpp.execution.ExecutionResult;
+import org.apache.iotdb.db.mpp.execution.IQueryExecution;
+import org.apache.iotdb.db.mpp.execution.QueryStateMachine;
 import org.apache.iotdb.db.mpp.sql.statement.Statement;
 import org.apache.iotdb.rpc.RpcUtils;
 import org.apache.iotdb.rpc.TSStatusCode;
@@ -105,6 +108,11 @@ public class ConfigExecution implements IQueryExecution {
   // Generate the corresponding IConfigTask by statement.
   // Each type of statement will has a ConfigTask
   private IConfigTask getTask(Statement statement) {
-    throw new NotImplementedException();
+    switch (statement.getType()) {
+      case SET_STORAGE_GROUP:
+        return new SetStorageGroupTask(statement);
+      default:
+        throw new NotImplementedException();
+    }
   }
 }
