+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/config/SampleConfigTask.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.execution;
+package org.apache.iotdb.db.mpp.execution.config;
 
 import org.apache.iotdb.db.mpp.sql.statement.Statement;
