+++ server/src/main/java/org/apache/iotdb/db/mpp/sql/analyze/Analyzer.java
@@ -46,6 +46,7 @@ import org.apache.iotdb.db.mpp.sql.statement.crud.*;
 import org.apache.iotdb.db.mpp.sql.statement.metadata.AlterTimeSeriesStatement;
 import org.apache.iotdb.db.mpp.sql.statement.metadata.CreateAlignedTimeSeriesStatement;
 import org.apache.iotdb.db.mpp.sql.statement.metadata.CreateTimeSeriesStatement;
+import org.apache.iotdb.db.mpp.sql.statement.metadata.SetStorageGroupStatement;
 import org.apache.iotdb.db.mpp.sql.statement.sys.AuthorStatement;
 import org.apache.iotdb.db.qp.constant.SQLConstant;
 import org.apache.iotdb.tsfile.exception.filter.QueryFilterOptimizationException;
@@ -224,6 +225,16 @@ public class Analyzer {
 
       String devicePath = alterTimeSeriesStatement.getPath().getDevice();
       analysis.setSchemaPartitionInfo(partitionFetcher.fetchSchemaPartitionInfo(devicePath));
+      return analysis;
+    }
+
+    @Override
+    public Analysis visitSetStorageGroup(
+        SetStorageGroupStatement setStorageGroupStatement, MPPQueryContext context) {
+      Analysis analysis = new Analysis();
+      analysis.setStatement(setStorageGroupStatement);
+
+      // TODO SpriCoder check what to do
       return analysis;
     }
